requirements.txt
aiohttp
aiohttp-retry
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
cloudinary
dj-database-url
Django
django-cloudinary-storage
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
frozenlist
idna
inflection
multidict
packaging
Pillow
psycopg2
PyJWT
python-decouple
pytz
PyYAML
requests
six
sqlparse
twilio
typing_extensions
uritemplate
urllib3
yarl